// src/features/gas/types.ts
export interface GasList {
  compressor: number;
  daily: number;
  flow_decanting: number;
  habis: number;
  id: string;
  jam: string;
  jtb_akumulasi: number;
  jtb_flow: number;
  jtb_serap: number;
  jtb_tekanan: number;
  kjg_akumulasi: number;
  kjg_flow: number;
  kjg_serap: number;
  kjg_tekanan: number;
  spp_flow: number;
  spp_tekanan: number;
  stock: number;
  tanggal: string;
}